feat : 뉴스피드 인증 추가

+++ src/main/java/com/sparta/oneandzerobest/newsfeed/service/NewsfeedService.java
@@ -1,5 +1,8 @@
 package com.sparta.oneandzerobest.newsfeed.service;
 
+import com.sparta.oneandzerobest.auth.entity.User;
+import com.sparta.oneandzerobest.auth.repository.UserRepository;
+import com.sparta.oneandzerobest.auth.util.JwtUtil;
 import com.sparta.oneandzerobest.newsfeed.dto.NewsfeedRequestDto;
 import com.sparta.oneandzerobest.newsfeed.dto.NewsfeedResponseDto;
 import com.sparta.oneandzerobest.newsfeed.entity.Newsfeed;
@@ -20,10 +23,16 @@ import org.springframework.web.bind.annotation.RestController;
 public class NewsfeedService {
 
     private final NewsfeedRepository newsfeedRepository;
+    private final JwtUtil jwtUtil;
+    private final UserRepository userRepository;
 
-    public ResponseEntity<NewsfeedResponseDto> postContent(NewsfeedRequestDto contentRequestDto) {
+    public ResponseEntity<NewsfeedResponseDto> postContent(String token, NewsfeedRequestDto contentRequestDto) {
 
         try {
+            if (!validToken(token)) {
+                throw new IllegalArgumentException("잘못된 토큰입니다");
+            }
+            String userid = jwtUtil.getUsernameFromToken(token);
             Newsfeed newsfeed = new Newsfeed(contentRequestDto.getUserid(), contentRequestDto.getContent());
             newsfeedRepository.save(newsfeed);
 
@@ -55,10 +64,14 @@ public class NewsfeedService {
     }
 
     @Transactional
-    public ResponseEntity<NewsfeedResponseDto> putContent(Long contentId,
+    public ResponseEntity<NewsfeedResponseDto> putContent(String token, Long contentId,
         NewsfeedRequestDto contentRequestDto) {
 
         try {
+            if (!validToken(token)) {
+                throw new IllegalArgumentException("잘못된 토큰입니다");
+            }
+
             Newsfeed newsfeed = newsfeedRepository.findById(contentId).orElseThrow(() -> new RuntimeException("Content not found"));
 
             newsfeed.setContent(contentRequestDto.getContent());
@@ -71,9 +84,13 @@ public class NewsfeedService {
     }
 
 
-    public ResponseEntity<Long> deleteContent(Long contentId) {
+    public ResponseEntity<Long> deleteContent(String token, Long contentId) {
 
         try{
+            if (!validToken(token)) {
+                throw new IllegalArgumentException("잘못된 토큰입니다");
+            }
+
             Newsfeed content = newsfeedRepository.findById(contentId).orElseThrow(() -> new RuntimeException("Content not found"));
             newsfeedRepository.delete(content);
             return ResponseEntity.ok(content.getId());
@@ -81,5 +98,10 @@ public class NewsfeedService {
         catch(RuntimeException e){
             return ResponseEntity.badRequest().header("잘못된 요청입니다.").body(null);
         }
+    }
+
+    private User validToken(String token) {
+        String username = jwtUtil.getUsernameFromToken(token.replace("Bearer ", ""));
+        return userRepository.findByUsername(username).orElse(null);
     }
 }
